+++ client/verta/tests/conftest.py
@@ -364,6 +364,15 @@ def model_version(registered_model):
 
 
 @pytest.fixture
+def endpoint(client, created_endpoints):
+    path = _utils.generate_default_name()
+    endpoint = client.create_endpoint(path)
+    created_endpoints.append(endpoint)
+
+    yield endpoint
+
+
+@pytest.fixture
 def created_endpoints(client):
     to_delete = []
 
@@ -382,7 +391,7 @@ def organization(client):
 
     org.delete()
 
-    
+
 @pytest.fixture
 def requirements_file():
     with tempfile.NamedTemporaryFile('w+') as tempf:
@@ -395,4 +404,3 @@ def requirements_file():
         tempf.seek(0)
 
         yield tempf
-