+++ client/verta/verta/_cli/deployment/update.py
@@ -7,7 +7,7 @@ from .deployment import deployment
 from ... import Client
 from ...endpoint.update._strategies import DirectUpdateStrategy, CanaryUpdateStrategy
 from ...endpoint.update.rules import _UpdateRule
-from ...endpoint.resources import _Resource
+from ...endpoint.resources import Resources
 from ...endpoint.autoscaling import Autoscaling
 from ...endpoint.autoscaling.metrics import _AutoscalingMetric
 from ..._registry import RegisteredModelVersion
@@ -93,9 +93,9 @@ def update_endpoint(path, run_id, model_version_id, filename, strategy, resource
             strategy_obj.add_rule(_UpdateRule._from_dict(json.loads(rule)))
 
     if resources:
-        resources_list = _Resource._from_dict(json.loads(resources))
+        resources = Resources._from_dict(json.loads(resources))
     else:
-        resources_list = None
+        resources = None
 
     if autoscaling:
         autoscaling_obj = Autoscaling._from_dict(json.loads(autoscaling))
@@ -109,4 +109,4 @@ def update_endpoint(path, run_id, model_version_id, filename, strategy, resource
     else:
         env_vars_dict = None
 
-    endpoint.update(model_reference, strategy_obj, resources=resources_list, autoscaling=autoscaling_obj, env_vars=env_vars_dict)
+    endpoint.update(model_reference, strategy_obj, resources=resources, autoscaling=autoscaling_obj, env_vars=env_vars_dict)