+++ client/verta/docs/changelog.rst
@@ -4,7 +4,7 @@ Release Notes
 
 .. This comment block is a template for version release notes.
    v.. (--)
-   ---------------------
+   --------------------
 
    Backwards Incompatibilities
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -35,6 +35,46 @@ Release Notes
    ^^^^^^^^^^^^^^^^
    - `
      <>`__
+
+
+v0.15.2 (2020-08-28)
+--------------------
+
+Enhancements
+^^^^^^^^^^^^
+- `add retries for all client connection errors
+  <https://github.com/VertaAI/modeldb/pull/1407>`__
+
+
+v0.15.1 (2020-08-24)
+--------------------
+
+Bug Fixes
+^^^^^^^^^
+- `correctly log model artifacts attribute to enable custom models with dependencies
+  <https://github.com/VertaAI/modeldb/pull/1399>`__
+
+
+v0.15.0 (2020-08-24)
+--------------------
+
+New Features
+^^^^^^^^^^^^
+- `Verta model registry
+  <https://docs.verta.ai/en/master/api/api/registered_model.html>`__
+- `Verta endpoints
+  <https://docs.verta.ai/en/master/api/api/endpoint.html>`__
+
+Enhancements
+^^^^^^^^^^^^
+- `expand custom modules virtual environment filter
+  <https://github.com/VertaAI/modeldb/pull/1392>`__
+
+Bug Fixes
+^^^^^^^^^
+- `fix bug where multipart upload loop may have an undefined variable in the
+  event of connection errors
+  <https://github.com/VertaAI/modeldb/pull/1362>`__
 
 
 v0.14.17 (2020-08-13)